+++ DNA Sequence Transformations/timingJsonToCSV.c
@@ -96,7 +96,7 @@ int timingJsonToCSV(char * input, FILE * ofp, int maxNumInputs, int m, int k){
 			numBuf[a++] = c;
 		}
 		numBuf[a] = '\x00';
-		timeArr[i] = atoi(numBuf); 
+		timeArr[i] = atof(numBuf); 
 		end += timingMatch.rm_eo;
 	}
